+++ test/support/fake_query.dart
@@ -29,7 +29,7 @@ class FakeQuery extends KoolbaseQuery {
   int listens = 0;
 
   @override
-  Future<QueryResult> get() => onGet();
+  Future<QueryResult> get({bool fresh = false}) => onGet();
 
   @override
   Stream<QueryResult> get stream {